增加生成exe方法

+++ PBSpyORCA/Program.cs
@@ -9,6 +9,7 @@ namespace PBSpyORCA
         {
             for (int i = 0; i < 10; i++)
             {
+                string runPath = AppContext.BaseDirectory;
                 string lib = "test.pbl";
                 string app = "test";
                 string comments = "你好Hucxy";
@@ -46,23 +47,8 @@ namespace PBSpyORCA
                     Console.WriteLine("设置当前应用失败");
                     return;
                 }
-                PBORCA_ENTRYINFO appInfo = new PBORCA_ENTRYINFO();
-                ret = PBORCA_LibraryEntryInformation(session, lib, app, PBORCA_TYPE.PBORCA_APPLICATION, ref appInfo);
-                if (ret != 0)
-                {
-                    Console.WriteLine("获取对象信息失败");
-                    return;
-                }
-                StringBuilder sbSra = new StringBuilder(appInfo.lSourceSize + 2);
-                ret = PBORCA_LibraryEntryExport(session, lib, app, PBORCA_TYPE.PBORCA_APPLICATION, sbSra, appInfo.lSourceSize + 2);
-                if (ret != 0)
-                {
-                    Console.WriteLine("导出对象失败");
-                    return;
-                }
-                sra = sbSra.ToString();
 
-                StringBuilder sbComments = new StringBuilder(256);
+                var sbComments = new StringBuilder(256);
                 ret = PBORCA_LibraryDirectory(session, lib, sbComments, 512, Marshal.GetFunctionPointerForDelegate<PBORCA_CallBack>(CallBack), IntPtr.Zero);
                 if (ret != 0)
                 {
@@ -71,11 +57,61 @@ namespace PBSpyORCA
                 }
                 comments = sbComments.ToString();
 
+                var appInfo = new PBORCA_ENTRYINFO();
+                ret = PBORCA_LibraryEntryInformation(session, lib, app, PBORCA_TYPE.PBORCA_APPLICATION, ref appInfo);
+                if (ret != 0)
+                {
+                    Console.WriteLine("获取对象信息失败");
+                    return;
+                }
+
+                var sbSra = new StringBuilder(appInfo.lSourceSize + 2);
+                ret = PBORCA_LibraryEntryExport(session, lib, app, PBORCA_TYPE.PBORCA_APPLICATION, sbSra, appInfo.lSourceSize + 2);
+                if (ret != 0)
+                {
+                    Console.WriteLine("导出对象失败");
+                    return;
+                }
+                sra = sbSra.ToString();
+
+                var exeInfo = new PBORCA_EXEINFO()
+                {
+                    lpszCompanyName = "lpszCompanyName",
+                    lpszProductName = "lpszProductName",
+                    lpszDescription = "lpszDescription",
+                    lpszCopyright = "lpszCopyright",
+                    lpszFileVersion = "1.0.0.0",
+                    lpszFileVersionNum = "2,0,0,0",
+                    lpszProductVersion = "3.0.0.0",
+                    lpszProductVersionNum = "4,0,0,0",
+                };
+                ret = PBORCA_SetExeInfo(session, ref exeInfo);
+                if (ret != 0)
+                {
+                    Console.WriteLine("设置编译信息失败");
+                    return;
+                }
+
+                string lpszExeName = Path.Combine(runPath, "test.exe");
+                ret = PBORCA_ExecutableCreate(session, lpszExeName, null, null, IntPtr.Zero, IntPtr.Zero, [0], 1, 0);
+                if (ret != 0)
+                {
+                    Console.WriteLine("创建可执行文件失败");
+                    return;
+                }
+
                 PBORCA_SessionClose(session);
                 File.Delete(lib);
             }
             Console.ReadKey();
         }
+
+        static void CallBack(IntPtr a1, IntPtr pUserData)
+        {
+            var stru = Marshal.PtrToStructure<PBORCA_DIRENTRY>(a1);
+            Console.WriteLine(string.Format("{0}\t{1}\t{2}\t{3}\t{4}", stru.lpszEntryName, stru.lEntrySize, stru.lCreateTime, stru.otEntryType, Encoding.Unicode.GetString(stru.szComments)));
+        }
+
         internal enum PBORCA_TYPE
         {
             PBORCA_APPLICATION,
@@ -111,15 +147,41 @@ namespace PBSpyORCA
             public int lObjectSize;
             public int lSourceSize;
         }
+        [StructLayout(LayoutKind.Sequential, Pack = 1)]
+        public struct PBORCA_EXEINFO
+        {
+            [MarshalAs(UnmanagedType.LPTStr)]
+            public string lpszCompanyName;
+            [MarshalAs(UnmanagedType.LPTStr)]
+            public string lpszProductName;
+            [MarshalAs(UnmanagedType.LPTStr)]
+            public string lpszDescription;
+            [MarshalAs(UnmanagedType.LPTStr)]
+            public string lpszCopyright;
+            [MarshalAs(UnmanagedType.LPTStr)]
+            public string lpszFileVersion;
+            [MarshalAs(UnmanagedType.LPTStr)]
+            public string lpszFileVersionNum;
+            [MarshalAs(UnmanagedType.LPTStr)]
+            public string lpszProductVersion;
+            [MarshalAs(UnmanagedType.LPTStr)]
+            public string lpszProductVersionNum;
+        }
 
-        [DllImport("PBSpy.dll", CharSet = CharSet.Unicode)]
-        public static extern void PBORCA_SessionClose(IntPtr hORCASession);
+        public delegate void PBORCA_CallBack(IntPtr a1, IntPtr a2);
+
+        #region managing the ORCA session
         [DllImport("PBSpy.dll", CharSet = CharSet.Unicode)]
         public static extern IntPtr PBORCA_SessionOpen(int pbVer);
+        [DllImport("PBSpy.dll", CharSet = CharSet.Unicode)]
+        public static extern void PBORCA_SessionClose(IntPtr hORCASession);
         [DllImport("PBSpy.dll", CharSet = CharSet.Unicode)]
         public static extern int PBORCA_SessionSetLibraryList(IntPtr hORCASession, string[] pLibNames, int iNumberOfLibs);
         [DllImport("PBSpy.dll", CharSet = CharSet.Unicode)]
         public static extern int PBORCA_SessionSetCurrentAppl(IntPtr hORCASession, string lpszApplLibName, string lpszApplName);
+        #endregion
+
+        #region managing PowerBuilder libraries
         [DllImport("PBSpy.dll", CharSet = CharSet.Unicode)]
         public static extern int PBORCA_LibraryCreate(IntPtr hORCASession, string lpszLibraryName, string lpszLibComments);
         [DllImport("PBSpy.dll", CharSet = CharSet.Unicode)]
@@ -128,18 +190,20 @@ namespace PBSpyORCA
         public static extern int PBORCA_LibraryEntryInformation(IntPtr hORCASession, string lpszLibraryName, string lpszEntryName, PBORCA_TYPE otEntryType, ref PBORCA_ENTRYINFO pEntryInformationBlock);
         [DllImport("PBSpy.dll", CharSet = CharSet.Unicode)]
         public static extern int PBORCA_LibraryEntryExport(IntPtr hORCASession, string lpszLibraryName, string lpszEntryName, PBORCA_TYPE otEntryType, StringBuilder lpszExportBuffer, int lExportBufferSize);
+        #endregion
+
+        #region importing and compiling PowerBuilder objects
         [DllImport("PBSpy.dll", CharSet = CharSet.Unicode)]
         public static extern int PBORCA_CompileEntryImport(IntPtr hORCASession, string lpszLibraryName, string lpszEntryName, PBORCA_TYPE otEntryType, string lpszComments, string lpszEntrySyntax, int lEntrySyntaxBuffSize, IntPtr pCompErrorProc, IntPtr pUserData);
         [DllImport("PBSpy.dll", CharSet = CharSet.Unicode)]
         public static extern int PBORCA_CompileEntryImportList(IntPtr hORCASession, string[] pLibraryNames, string[] pEntryNames, PBORCA_TYPE[] otEntryTypes, string[] pComments, string[] pEntrySyntaxBuffers, int[] pEntrySyntaxBuffSizes, int iNumberOfEntries, IntPtr pCompErrorProc, IntPtr pUserData);
+        #endregion
 
-        public delegate void PBORCA_CallBack(IntPtr a1, IntPtr a2);
-
-
-        static void CallBack(IntPtr a1, IntPtr pUserData)
-        {
-            var stru = Marshal.PtrToStructure<PBORCA_DIRENTRY>(a1);
-            Console.WriteLine(string.Format("{0}\t{1}\t{2}\t{3}\t{4}", stru.lpszEntryName, stru.lEntrySize, stru.lCreateTime, stru.otEntryType, Encoding.Unicode.GetString(stru.szComments)));
-        }
+        #region creating executables and dynamic libraries
+        [DllImport("PBSpy.dll", CharSet = CharSet.Unicode)]
+        public static extern int PBORCA_SetExeInfo(IntPtr hORCASession, ref PBORCA_EXEINFO pExeInfo);
+        [DllImport("PBSpy.dll", CharSet = CharSet.Unicode)]
+        public static extern int PBORCA_ExecutableCreate(IntPtr hORCASession, string lpszExeName, string? lpszIconName, string? lpszPBRName, IntPtr pLinkErrProc, IntPtr pUserData, int[] iPBDFlags, int iNumberOfPBDFlags, int lFlags);
+        #endregion
     }
 }
